Add support for knockouts
This patch adds scheduling and scoring support for the knockouts.
These matches are added to the existing match schedule list, but are
marked as knockout matches.  This allows all systems to treat them in
the same manner as league matches if they want.

Since the teams in a match in the knockouts depend on the matches that
came before, matches where the players are unknown are filled with
"???" instead of a real team name.  This means that things like the
match schedule will display "???" vs "???" etc with the correct match
time.  Once the players are known, these are filled out correctly.

So, initially, the knockout schedule will entirely consist of matches
of "???" playing three other "???" players.

Scores for the knockouts will be stored in the /knockouts directory of
the compstate repository in the same manner as the league matches.
(The match numbering continues.)

Knockout rounds are scheduled with a configurable time gap between
them.  This allows time for score entry.  The final round can have an
additional gap too, as that one is the most stressful for everyone.

This currently lacks two (non-critical) things:  (both coming soon)
 - Support for delays during the knockouts
 - A stable random shuffling situation
   (The teams are placed in random zones during the knockout
    scheduling.  Since this algorithm is run on multiple systems, etc,
    the random sorting needs to be stable across systems and python
    versions.  At the moment this implementation contains a 'random'
    implementation that doesn't actually do anything.)