Use UTC internally
Using UTC internally means the system will be able to cope if it runs
over a DST change, but also that the HTTP API will return an ISO8601
date in UTC rather than in local time, which causes the clients to have
to cope with the timezone themselves and wouldn't work if they are
working in a different timezone to the server anyway.

This change will need to be reflected in srcomp-http (switch to utcnow)
and in the compstate repo (the timezone should be specified in the dates
where they are not UTC, which, at the time of writing, is BST).