Separate construction of KnockoutScheduler from MatchSchedule.
This makes it possible to interact with the latter on its own if
desired (such as in tests), and makes it much simpler to get most
of the tests working again. Provide a class method which can be
used in place of the constructor (with the same signature as it
had) and use this for the 'real' usage of the object.

Change-Id: I76bafcbb07d2d63437a42b928d3bcb9036f417d0