Avoid a KeyError when an arena is unused.
Previously, if you requested the match_after a given time for an
arena not in some matches, you got a KeyError. While that's somewhat
valid for arenas which don't exist, it's not valid for the finals
where arena B is going to be unused.

Change to a safe lookup of the arena in the match doublet and also
change the match_at method to use the same code style. This makes
it clearer that the check is needed I hope.

Change-Id: I3b4a45a0f46c877a2b4eac96d1673d242709f921