Extend timing checks to catch overlaps too.
The previous check only considered the case where matches erronously
started at the same time. Extend that support to take into account
the fact that the matches have duration. Update the dummy submodule
to correct for it having the knockouts at a time which would conflict
with the last league match.

Change-Id: I01487dddcdb9851d4d14a98952799d87ec289878