Make the yaml parser return timezone aware datetimes
By default, PyYAML returns datetime objects that are timezone naive,
but are actually UTC.  This isn't good, because we're in BST at the
moment.

This patch overrides the interpretation of the timestamp nodes from
the YAML so that timezone aware objects are returned.  This fixes a
bunch of issues involving the arena screens.

Change-Id: I408a84b079b3fc375dde0bc07e2b180616afc543