Extract the scorer to a dependency injection.
The details of the scorer are going to change each year. Even if it
is always in the same place for usage, we really want to be able
to test things without being dependent on the scorer. By extracting
it to an injected dependency we make changing it very much easier.

Change-Id: I667af5494cb3568a8df8a9b51889f45af6e409b8