Change to loading the scorer dynamically.
While for the moment the values are all literals, this provides a
possible mechanism for the scorer class to be both located and then
loaded in a dynamic manner. This would make it possible to have the
scorer implementation specified as a config option, or even as part
of the competition state.

Change-Id: Iadb6e36ae5d641553bf7757699142737454f2e08