Fix #2410: Correct handling of multiple delays.
Previously, having multiple delays in the same period would result
in the delay being applied only to matches whose _original_ start
time was within the range of the delay, rather than the matches
actually starting after it should apply. This resulted in rather
confusing results, and made it rather hard to apply delays in the
real world.

Change-Id: I7f5e2b03f56a4c0063f37434390996af8dd68b71