Add a calculation of the league positions.
This adds a new positions property to the LeagueScores class that
returns a dictionary of tla -> position. This is calculated from
the current league scores and correctly handles league-point ties.

At the moment, this introduces a TODO on how we deal with ties of
both league points and game points. As far as I'm aware this hasn't
actually ever happened, but it is possible.

Change-Id: Ic112bcb6fe954cda7d9280049322a3a6bd6c6a1d