Fix winner selection in static knockouts
I'd not noticed that the ranking returned by get_ranking is ordered
in reverse compared to what would usually be expected (ie the winner
is at the end of the list). Reverse the list before looking up the
team at a given position to give the right behaviour. Add tests for
various stages through the knockouts which showed this bug.

Change-Id: Ic63dbed07a4462592d8ff8e65d0f8735e1ffe0d3