Restore the previous testing for matches_at
This was removed along with the previous implementation, but wasn't
restored with the replacement. Re-work it for the slightly different
output format and replace the current testing with it.

Fix a bug this showed up where it was possible to be 'in' two matches
at the same time in the same arena at the exact end/start changeover
time. Resolve this in the same way as before by not claiming to be
'in' a given match at its end time.

Change-Id: Ic23027ce55be5724aca59329be0172b4e7c9ebb2