Remove duplicate use of the term 'period'
This originally meant a stretch of time during which matches would
be run, but appears to have also been used to mean the duration of
a match. Since then the latter appears to have been epanded upon,
causing confusion. Replace the with 'duration' where suitable or
with 'slot' when refering to the specific timeslot a match occurs
within.

Change-Id: I290c8b360878035df18387137c72e9d998bfcc7e