Keep positions data around on all scored matches
Previously, positions data (the mapping of 1st->[teams] etc) was
calculated within ranker for each match, but only as an intermediate
result to compute the ranked points.

For awards detection and for the tiebreaker, this meant recomputing the
positions for the finals.

Instead, keep the positions data around on all matches.