Fix #2826: Extract delay handling to a separate class
This creates a new MatchPeriodClock class which handles the passage
of time when doing scheduling. It is aware of the applicable delays
as well as the end and max-end times of the period it is scheduling
for. It deliberately makes no assumptions as to how time is being
measured and knows nothing about the matches themselves.

Use this new clock class to replace the timing handling in both the
league and knockout schedulers, thus ensuring that time and delay
handling is consistent throughout SRComp.

Change-Id: I060076ae1c552c9269babcfa8b6bb03481d02a63