Revert the sub-classing of Random in stable_random
It turns out that this resulted in different shuffle orderings for
lists under Python 3 than Python 2. Since shuffle is used in the
knockouts, in turn this lead to different corner placements.

This could have lead to much confusion at the competition since
different sets of output (venue displays, shepherd data, website)
are expecting to be using different Python versions.

Change-Id: I244bc5be243b1e7f1e27bc054c0840148b8fb57d