Make stage's paths work relative to the compstate
This makes more sense than it being relative to the current directory
since all the operations here are on the compstate.

Update the get_score_path function to return an absolute path to avoid
that causing any issues. This is comment suggests it should have been
doing anyway, but wasn't.

Change-Id: Iea09e972190555b482e8ae5f78718be34bc336b8