Add a mechanism for spacing out matches
For SR2015 we're going to have some more space between the first
few matches in order to let the linesmen, shepherds and teams get
used to doing actual matches. We therefore need a way to do this
spacing which won't appear as a 'delay' value (since it's planned
ahead of time).

Introduce a new 'league' key in the schedule which has a 'extra_spacing'
key which is a list of such spacings. Use this to add spacing as
needed in front of the given matches (while still allowing for
delays to further adjust the timings).

Change-Id: Id64faa2e9344cfa36d79bd8ec92b4bbe51b1d2c1