Fix #2952: Make awards calculations aware of tiebreakers
This now looks at the tiebreaker first if there is one, and includes
fall-back to the final if one of the teams was knocked out at that
point. It assumes that the final is the match immediately preceding
the tiebreaker and occured in the same arena. This feels like a safe
assumption for now, but clearly might change in the future or if
several tiebreakers are played.

Change-Id: I6c81de099129373ed28cc4b62e06a9467cec8f51