Move knockout ranking into score handling
This creates a new attribute on the 'KnockoutScores' class which
stores the resolved positions from knockout matches. The knockout
scheduler is changed to use these data instead of its own calcuations,
the code for the new one having been largely copied from that.

Introduce new tests specifically for these resolved positions,
which are notably not in the same format as the 'game_positions'
attribute which already exists.

This works towards a resolution for #2977, though that will require
a small change in the HTTP API to use this new map as the source of
its data.

Change-Id: Id3b386219bf20f5796ff3da9fd618086f11b0752