Fix #3036: Fix dropout handling at start of knockouts
This resolves an off-by-error in the checking of the dropouts at
the start of the knockouts. Note that it still only handles teams
which have dropped out before the _start_ of the knockouts, and
doesn't cope with any form of dropout during the knockouts (whether
that is during the first round of knockouts or otherwise).

Change-Id: I59362a30d707b72df2f1ea4b01022cf55efb66ae