Infer and validate num-teams-per-arena from num-corners
This add support for arenas which have a number of zones which is
not four. We now infer the desired number of teams per match from
the number of corners in the arena spec and use that to validate
both the construction of each of the league, knockout and tiebreaker
matches.

Change-Id: Ief237405e4bde9ed041dfe11d54a29f4124abc97